requirements.txt
flask
flask_login
pymongo
flask_pymongo
flask_socketio
flask-user